Fix URLs to NNDC
URLs used to contain endf/b7.1, which has been changed to endf-b7.1. Files are still the same.

NNDC are also supporting https now, which they did not do previously. We are making the switch too. We don't need it from a security perspective, but I expect they are more likely to disable http at some point.

Closes #2.